# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ packages/sdk/src/Sdk.ts
@@ -16,9 +16,7 @@ import {
 import merge from 'lodash/merge';
 import './service/windowEventListener'; // to initial window events listener
 
-import {
-  Api, HandleByGlip, HandleByRingCentral, HandleByUpload,
-} from './api';
+import { Api, HandleByGlip, HandleByRingCentral, HandleByUpload } from './api';
 import { defaultConfig as defaultApiConfig } from './api/defaultConfig';
 import { AutoAuthenticator } from './authenticator/AutoAuthenticator';
 import DaoManager from './dao/DaoManager';
@@ -29,13 +27,14 @@ import notificationCenter from './service/notificationCenter';
 import { SyncService } from './module/sync';
 import { ApiConfig, DBConfig, ISdkConfig } from './types';
 import { AccountService } from './module/account';
-import { UserConfigService } from './module/config';
 import { setGlipToken } from './authenticator/utils';
 import { AccountGlobalConfig } from './module/account/config';
 import { ServiceConfig, ServiceLoader } from './module/serviceLoader';
 import { PhoneParserUtility } from './utils/phoneParser';
 import { configMigrator } from './framework/config';
 import { ACCOUNT_TYPE_ENUM } from './authenticator/constants';
+import { PermissionService, LaunchDarklyController, SplitIOController } from 'sdk/module/permission';
+
 
 const LOG_TAG = 'SDK';
 const AM = AccountManager;
@@ -54,6 +53,7 @@ class Sdk {
     public serviceManager: ServiceManager,
     public networkManager: NetworkManager,
     public syncService: SyncService,
+    public permissionService: PermissionService,
   ) {}
 
   async init(config: ISdkConfig) {
@@ -79,7 +79,9 @@ class Sdk {
     );
 
     if (!loginResp || !loginResp.success) {
-      window.indexedDB && window.indexedDB.deleteDatabase('Glip');
+      if (process.env.NODE_ENV !== 'test') {
+        window.indexedDB && window.indexedDB.deleteDatabase('Glip');
+      }
     }
     this._subscribeNotification();
     this._initDataAnalysis();
@@ -101,6 +103,9 @@ class Sdk {
     });
     Api.init(apiConfig, this.networkManager);
     await this.daoManager.initDatabase();
+
+    this.permissionService.injectControllers(new LaunchDarklyController());
+    this.permissionService.injectControllers(new SplitIOController());
 
     // Sync service should always start before login
     this.serviceManager.startService(SyncService.name);
@@ -196,9 +201,6 @@ class Sdk {
     this.networkManager.clearToken();
     this.serviceManager.stopAllServices();
     this.daoManager.deleteDatabase();
-    ServiceLoader.getInstance<UserConfigService>(
-      ServiceConfig.USER_CONFIG_SERVICE,
-    ).clear();
     AccountGlobalConfig.removeUserDictionary();
     this._resetDataAnalysis();
   }
